proyecto terminado

+++ proyecto-web/database/migrations/2014_10_12_000000_create_users_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->string('email')->unique();
             $table->timestamp('email_verified_at')->nullable();
             $table->string('password');
-            $table->integer('telefono')->nullable(); 
+            $table->bigInteger('telefono')->nullable(); 
             $table->string('direccion')->nullable();
             $table->string('ciudad')->nullable(); 
             $table->string('estado')->nullable();
